Tests/LibWeb/Text/expected/canvas/offscreencanvas-basic.txt

{"w":10,"h":10}
20px SerenitySans
20px SerenitySans
